Fixes and improvements
- Removed old place system
- Fixed an issue with clearing cache.
- Remove unnesseccary search check (combined both)
- Attempted to fix test_Switch.py (failed)
- Added comments in Functions
- Changed search.__FindFile to use glob.
- Remove correct file check from search
- Removed List option from search
- Updated file check to include hidden files (Except for .git with it's 255 folders)

+++ Program/Files/newSave.py
@@ -310,7 +310,10 @@ class save:
     @staticmethod
     def Delete(path):
         if os.path.exists(path):
-            shutil.rmtree(path)
+            if os.path.isdir(path):
+                shutil.rmtree(path)
+                return True
+            os.remove(path)
             return True
         print('Path not found! -> {}'.format(path))
         return False
